Move web config to app config (#245)
