--- client/wfnews-war/src/main/java/ca/bc/gov/nrs/wfnews/spring/WebConfig.java
@@ -1,20 +1,0 @@
-package ca.bc.gov.nrs.wfnews.spring;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.servlet.config.annotation.CorsRegistry;
-import org.springframework.web.servlet.config.annotation.EnableWebMvc;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-@Configuration
-@EnableWebMvc
-public class WebConfig implements WebMvcConfigurer {
-
-    @Override
-    public void addCorsMappings(CorsRegistry registry) {
-        registry.addMapping("/**")
-        .allowCredentials(true)
-        .allowedOriginPatterns("*")
-        .allowedHeaders("*")
-        .allowedMethods("GET", "POST", "OPTIONS");
-    }
-}
